Extract public hooks tests into a separate test suite file
